refactor(discord-guild-config): split into multiple services
