--- src/features/discord/guilds/constants/discord-guild-config.ts
@@ -1,6 +1,0 @@
-import { IDiscordGuildConfig } from "../../interfaces/discord-guild-config";
-
-export const DISCORD_GUILD_CONFIG: IDiscordGuildConfig = {
-  shouldWelcomeNewMembers: true,
-  soniaPermanentGuildInviteUrl: `https://discord.gg/PW4JSkv`,
-};
