+++ src/features/discord/guilds/enums/discord-guild-config-value-name.enum.ts
@@ -1,0 +1,4 @@
+export enum DiscordGuildConfigValueNameEnum {
+  SHOULD_WELCOME_NEW_MEMBERS = `welcome new members state`,
+  SONIA_PERMANENT_GUILD_INVITE_URL = `Sonia permanent guild invite url`,
+}
