+++ src/features/discord/messages/services/discord-message-error-service.ts
@@ -11,7 +11,7 @@ import { ellipsis } from "../../../../functions/formatters/ellipsis";
 import { GithubConfigService } from "../../../github/services/github-config-service";
 import { LoggerService } from "../../../logger/services/logger-service";
 import { DiscordChannelService } from "../../channels/services/discord-channel-service";
-import { DiscordGuildConfigService } from "../../guilds/services/discord-guild-config-service";
+import { DiscordGuildConfigService } from "../../guilds/services/config/discord-guild-config-service";
 import { DiscordSoniaService } from "../../users/services/discord-sonia-service";
 import { IDiscordMessageResponse } from "../interfaces/discord-message-response";
 import { AnyDiscordMessage } from "../types/any-discord-message";
